Refactor systems into separate files so it's a bit easier to find specific systems
